+++ CardSuit.cs
@@ -1,0 +1,15 @@
+﻿namespace PlayingCards
+{
+    /// <summary> Represents which suit a Card belongs to </summary>
+    public enum CardSuit 
+    { 
+        /// <summary> </summary>
+        Diamonds = 0,
+        /// <summary> </summary>
+        Hearts = 1,
+        /// <summary> </summary>
+        Spades = 2,
+        /// <summary> </summary>
+        Clubs = 3 
+    }
+}
