+++ DieType.cs
@@ -1,0 +1,21 @@
+﻿namespace Dice
+{
+    /// <summary>Represents the number of faces for a die</summary>
+    public enum DieType
+    {
+        /// <summary> Four Sided Die </summary>
+        D4 = 4,
+        /// <summary> Six Sided Die </summary>
+        D6 = 6,
+        /// <summary> Eight Sided Die </summary>
+        D8 = 8,
+        /// <summary> Ten Sided Die </summary>
+        D10 = 10,
+        /// <summary> Twelve Sided Die </summary>
+        D12 = 12,
+        /// <summary> Sixteen Sided Die </summary>
+        D16 = 16,
+        /// <summary> Twenty Sided Die </summary>
+        D20 = 20
+    }
+}
