+++ FaceValue.cs
@@ -1,0 +1,33 @@
+﻿namespace PlayingCards
+{
+    /// <summary> Represents the number/value of the card </summary>
+    public enum FaceValue 
+    {
+        /// <summary> </summary>
+        Ace = 1,
+        /// <summary> </summary>
+        Deuce = 2,
+        /// <summary> </summary>
+        Three = 3,
+        /// <summary> </summary>
+        Four = 4,
+        /// <summary> </summary>
+        Five = 5,
+        /// <summary> </summary>
+        Six = 6,
+        /// <summary> </summary>
+        Seven = 7,
+        /// <summary> </summary>
+        Eight = 8,
+        /// <summary> </summary>
+        Nine = 9,
+        /// <summary> </summary>
+        Ten = 10,
+        /// <summary> </summary>
+        Jack = 11,
+        /// <summary> </summary>
+        Queen = 12,
+        /// <summary> </summary>
+        King = 13 
+    }
+}
